requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
absl-py==2.0.0; python_version >= '3.7'
aiohttp==3.8.6; python_version >= '3.6'
aiosignal==1.3.1; python_version >= '3.7'
astor==0.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
async-timeout==4.0.3; python_version >= '3.7'
asynctest==0.13.0; python_version < '3.8'
attrs==23.1.0; python_version >= '3.7'
cachetools==5.3.2; python_version >= '3.7'
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.3.2; python_version >= '3.7'
cycler==0.11.0; python_version >= '3.6'
fonttools==4.38.0; python_version >= '3.7'
frozenlist==1.3.3; python_version >= '3.7'
gast==0.5.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-auth==2.23.4
grpcio==1.59.2; python_version >= '3.7'
h5py==3.8.0; python_version >= '3.7'
idna==3.4; python_version >= '3.5'
importlib-metadata==6.7.0; python_version < '3.10'
joblib==1.3.2; python_version >= '3.7'
keras-applications==1.0.8
keras-preprocessing==1.1.2
kiwisolver==1.4.5; python_version >= '3.7'
linkify-it-py==2.0.2
markdown-it-py[linkify,plugins]==2.2.0; python_version >= '3.7'
markdown==3.4.4; python_version >= '3.7'
markupsafe==2.1.3; python_version >= '3.7'
matplotlib==3.5.3
mdit-py-plugins==0.3.5
mdurl==0.1.2; python_version >= '3.7'
mock==5.1.0; python_version >= '3.6'
multidict==6.0.4; python_version >= '3.7'
numpy==1.21.6; python_version < '3.10' and platform_machine != 'aarch64' and platform_machine != 'arm64'
oauthlib==3.2.2
openai==0.28.1
packaging==23.2; python_version >= '3.7'
pandas==1.3.5
pillow==9.5.0; python_version >= '3.7'
protobuf==3.20.3
pyasn1-modules==0.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1==0.5.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pygments==2.16.1; python_version >= '3.7'
pyparsing==3.1.1; python_full_version >= '3.6.8'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2023.3.post1
requests==2.31.0; python_version >= '3.7'
rich==13.6.0; python_version >= '3.7'
rsa==4.9; python_version >= '3.6' and python_version < '4'
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
svg-to-gcode==1.5.4
svgwrite==1.4.3
tensorboard==1.13.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tensorflow-estimator==1.13.0
tensorflow==1.13.2
termcolor==2.3.0; python_version >= '3.7'
textual==0.41.0
threadpoolctl==3.1.0; python_version >= '3.6'
tqdm==4.66.1; python_version >= '3.7'
typing-extensions==4.7.1; python_version < '3.9'
uc-micro-py==1.0.2; python_version >= '3.7'
urllib3==2.0.7; python_version >= '3.7'
werkzeug==2.2.3; python_version >= '3.7'
wheel==0.41.3; python_version >= '3'
yarl==1.9.2; python_version >= '3.7'
zipp==3.15.0; python_version >= '3.7'
